Removes commented out prints for debug purposes from production code and removes
depricated debug printouts generated from lexer.go and parser.go.

Also fixes minor issues in template code for gofmt complience.

Fixes #53.