@sangisos 2017-07-28
Some benchmarking on https://github.com/llir/llvm:
```
gocc -zip ll.bnf

old lexer run 1
make  243.09s user 1.10s system 113% cpu 3:34.29 total

old lexer run 2
make  242.63s user 1.12s system 114% cpu 3:32.59 total

old lexer run 3
make  246.06s user 1.00s system 114% cpu 3:36.19 total

old lexer run 4
make  245.41s user 1.09s system 113% cpu 3:36.72 total

old lexer run 5
make 243.37s user 1.09s system 112% cpu 3:37.21 total


new lexer run 1
make  239.61s user 1.10s system 112% cpu 3:34.20 total

new lexer run 2
make  242.72s user 1.11s system 114% cpu 3:33.28 total

new lexer run 3
make  240.61s user 1.09s system 112% cpu 3:34.67 total

new lexer run 4
make &> /dev/null  245.67s user 1.04s system 114% cpu 3:34.66 total

new lexer run 5
make &> /dev/null  242.82s user 1.09s system 112% cpu 3:35.94 total
```