@mewmew 2017-07-28
@awalterschulze Feel free to merge if you agree with my LGTM.