@mewmew 2017-07-29
> @mewmew yes. That makes sense. This is why I wanted you to make the bench =). Thanks.

:)

Must say, I'm a bit jealous of your contrib stats for this PR, any diff removing more lines than adding is a good diff in my book.

```diff
+40
-257
```